+++ R/zzz.R
@@ -1,6 +1,6 @@
 #' @import data.table
 #' @import checkmate
-#' @importFrom utils head tail adist
+#' @importFrom utils head tail adist removeSource
 #' @importFrom stats setNames as.formula terms runif
 #' @importFrom R6 R6Class is.R6
 #' @importFrom digest digest